# cedar-resource-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Better Metadata - Better Science",
    "version" : "2.8.3",
    "title" : "CEDAR Resource Server API"
  },
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/command/inclusions-subgraph-preview" : {
      "post" : {
        "tags" : [ "Command" ],
        "summary" : "Preview the inclusion subgraph of an artifact",
        "description" : "Build a preview of the tree of artifacts affected by a change to the given artifact, without applying any update.",
        "operationId" : "previewInclusionSubgraph",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/command/inclusions-subgraph-update" : {
      "post" : {
        "tags" : [ "Command" ],
        "summary" : "Update the inclusion subgraph of an artifact",
        "description" : "Propagate a change to the given artifact across the tree of affected artifacts, updating each referencing artifact on the artifact server.",
        "operationId" : "updateInclusionSubgraph",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
...
    "/command/make-artifact-not-open" : {
      "post" : {
        "tags" : [ "Command", "OpenView" ],
        "summary" : "Make artifact not open",
        "description" : "Make artifact not open.",
        "operationId" : "makeArtifactNotOpen",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "idRequest",
          "description" : "Id of the artifact to make open",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/IdRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/command/make-artifact-open" : {
      "post" : {
        "tags" : [ "Command", "OpenView" ],
        "summary" : "Make artifact open",
        "description" : "Make artifact open.",
        "operationId" : "makeArtifactOpen",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "idRequest",
          "description" : "Id of the artifact to make open",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/IdRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/command/make-folder-not-open" : {
      "post" : {
        "tags" : [ "Command", "OpenView" ],
        "summary" : "Make folder not open",
        "description" : "Make folder not open.",
        "operationId" : "makeFolderNotOpen",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "idRequest",
          "description" : "Id of the folder to make open",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/IdRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/command/make-folder-open" : {
      "post" : {
        "tags" : [ "Command", "OpenView" ],
        "summary" : "Make folder open",
        "description" : "Make folder open.",
        "operationId" : "makeFolderOpen",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "idRequest",
          "description" : "Id of the folder to make open",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/IdRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "api_key" : {
      "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "IdRequest" : {
      "type" : "object",
      "properties" : {
        "@id" : {
          "type" : "string",
          "description" : "Unique URL identifier representing an artifact."
        }
      },
      "description" : "Id of the artifact."
    },
...
  }
}